add button to display zaken count
