+++ src/rma/destruction/templates/destruction/destructionlist_create.html
@@ -2,10 +2,6 @@
 
 {% block content %}
 
-<header>
-    <h1 class="title">Create new destruction list</h1>
-</header>
-
 <article id="content-main">
 
     {{ zaaktypen|json_script:"zaaktype-choices" }}
