+++ src/rma/js/components/destruction-list/zaken-table.js
@@ -1,4 +1,4 @@
-import React, {useEffect, useState} from 'react';
+import React, {useState} from 'react';
 import { CheckboxInput} from "./inputs";
 
 
@@ -6,45 +6,11 @@ function displayZaaktype (zaaktype) {
     return `${zaaktype.omschrijving} (${zaaktype.versiedatum}}`;
 }
 
-function getFullUrl(url, filters) {
-    const query = Object.keys(filters).filter(k=>filters[k]).map(k => `${k}=${filters[k]}`).join('&');
-    return `${url}?${query}`;
-}
 
 function ZakenTable(props) {
-    const { url, filters } = props;
+    const { zaken, isLoaded, error, checkboxes, setCheckboxes } = props;
 
-    //load zaken
-    const [error, setError] = useState(null);
-    const [isLoaded, setIsLoaded] = useState(false);
-    const [zaken, setZaken] = useState([]);
-
-    // checkboxes
     const [selectAll, setSelectAll] = useState(false);
-    const [checkboxes, setCheckboxes] = useState({});
-    console.log("checkboxes=", checkboxes);
-
-    useEffect(() => {
-        const fullUrl = getFullUrl(url, filters);
-        window.fetch(fullUrl)
-            .then(res => res.json())
-            .then(
-                (result) => {
-                    setIsLoaded(true);
-                    setZaken(result.zaken);
-
-                    // refresh checkboxes and deselect them
-                    const refreshedCheckboxes = result.zaken.reduce((result, zaak) => {
-                            return {...result, [zaak.url]: false};
-                        }, {});
-                    setCheckboxes(refreshedCheckboxes);
-                },
-                (error) => {
-                    setIsLoaded(true);
-                    setError(error);
-                }
-            )
-      }, [filters])
 
     if (error) {
         return <div>Error in fetching zaken: {error.message}</div>;
