+++ src/rma/sass/components/_destruction-create.scss
@@ -1,11 +1,27 @@
 .destruction-create {
-  display: flex;
+  &__header {
+    display: flex;
+    align-items: center;
+  }
 
-  & &__filters {
+  &__title {
+    flex: 1;
+  }
+
+  &__nav {
+    text-align: right;
+    margin-right: 0.3em;
+  }
+
+  &__content {
+    display: flex;
+  }
+
+  &__filters {
     width: 30%;
   }
 
-  & &__zaken {
+  &__zaken {
     width: 70%
   }
 }
